+++ CHANGELOG.md
@@ -77,3 +77,9 @@ npm install -g laravel-echo-server
 # 0.0.2
 
  - Send private channel auth data by event ("channel_subscribe_success").
+
+# 0.0.3
+
+  - Fix some bugs.
+  - Change welcome message.
+  - Add FirebaseAdmin service.